backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openai-whisper
pydub
soundfile
numpy
deep-translator==1.11.4
